+++ flutter/lib/ui/history/benchmark_export_result_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:mlperfbench/data/results/benchmark_result.dart';
 import 'package:mlperfbench/data/results/dataset_info.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
+import 'package:mlperfbench/ui/formatter.dart';
 import 'package:mlperfbench/ui/history/utils.dart';
-import 'package:mlperfbench/ui/time_utils.dart';
 
 class BenchmarkExportResultScreen extends StatefulWidget {
   final BenchmarkExportResult result;
